requirements.txt
async-timeout==5.0.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==308
redis==5.2.0
requests==2.32.3
SpeechRecognition==3.11.0
spotipy==2.24.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
